requirements.txt
# use pip install torch -f https://download.pytorch.org/whl/torch_stable.html to install
# pytorch
torch==1.8.0+cu111
torchvision==0.9.0+cu111
opencv-python~=4.5.3.56
numpy~=1.19.5
pyqt5~=5.15.4
pillow~=8.4.0